Refactor: Improve domain naming and decimal precision - Use BigDecimal for all financial variables and methods to eliminate precision loss - Refactor variable and method naming to align with financial terminology - Add constants for rounding scale and modes - Update README to reflect rebalancing logic

+++ src/test/java/com/accountabc/model/Security.java
@@ -2,11 +2,13 @@ package com.accountabc.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Security {
     private String name;
-    private int target;
-    private int current;
-    private int targetVariance;
-    private double unitPrice;
+    private BigDecimal target;
+    private BigDecimal current;
+    private BigDecimal targetVariance;
+    private BigDecimal unitPrice;
 }
